+++ OrionUO/Gumps/GumpPartyManifest.cpp
@@ -44,15 +44,15 @@ void CGumpPartyManifest::UpdateContent()
 
 	IFOR(i, 0, 10)
 	{
-		Add(new CGUIButton(ID_GPM_BUTTON_TELL_MEMBER + i, 0x0FAB, 0x0FAC, 0x0FAD, 40, yPtr + 2));
+		Add(new CGUIButton(ID_GPM_BUTTON_TELL_MEMBER + (int)i, 0x0FAB, 0x0FAC, 0x0FAD, 40, yPtr + 2));
 
 		if (isLeader)
-			Add(new CGUIButton(ID_GPM_BUTTON_KICK_MEMBER + i, 0x0FB1, 0x0FB2, 0x0FB3, 80, yPtr + 2));
+			Add(new CGUIButton(ID_GPM_BUTTON_KICK_MEMBER + (int)i, 0x0FB1, 0x0FB2, 0x0FB3, 80, yPtr + 2));
 
 		Add(new CGUIGumppic(0x0475, 130, yPtr));
 
 		text = (CGUIText*)Add(new CGUIText(0x0386, 140, yPtr + 1));
-		text->CreateTextureA(2, g_Party.Member[i].GetName(i + 1), 250, TS_CENTER);
+		text->CreateTextureA(2, g_Party.Member[i].GetName((int)i + 1), 250, TS_CENTER);
 
 		yPtr += 25;
 	}
